Add error message for missing font family name
Hey,
tried exporting a new design today and took the time to figure out why the script is failing to generate the font.

I simply hadn't set any font family names in my design, so the script crashes when it tries to copy over family names etc. This PR prevents the script from crashing, but instead will close the temporary copy and put a warning for the user to fix the issue.